+++ packages/runtime/tce-edit-vue3/src/components/ContentElementExample.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script setup lang="ts">
-import { defineProps, ref } from 'vue';
+import { ref } from 'vue';
 
 const props = defineProps<{
   id: string;