+++ packages/runtime/tce-edit-vue3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tailor-cms/tce-edit-next-runtime
 
+## 0.6.7
+
+### Patch Changes
+
+- Updated Embedded Container styling and btn props.
+
 ## 0.6.6
 
 ### Patch Changes